# RDK performance optimizations