--- a/.changeset/big-poems-design.md
+++ b/.changeset/big-poems-design.md
@@ -1,0 +1,5 @@
+---
+"@jspsych-contrib/plugin-rdk": minor
+---
+
+This version improves drawing performance through batching and simplified canvas clearing. It also introduces a new `dot_shape` parameter to allow drawing squares instead of circles, which can further reduce drawing time.