Hello,

These are a few changes to the RDK plugin to improve performance, particularly when there are more than a few hundred dots or on slower hardware.

 - Batched dot drawing. Rather than having a `beginPath()`/`fill()` pair per dot, wait to fill until the end.
 - Simplified dot clearing. All visuals seem to be redrawn each frame, so opted for `clearRect()` rather than drawing over individual dots.
 - Moved the fixation cue outside of `draw()`. Probably not a big deal, but it doesn't seem to be used per-aperture, and so would be unnecessarily redrawn for each aperture.
 - Introduced a `dot_shape` parameter, which allows squares to be substituted for circles in cases when dots are small. Squares can be drawn significantly faster.

For testing, I changed the number of dots in example4.html to `[500, 200, 100]`.

On my desktop (Ubuntu 20.04, RX 580 GPU, Chromium), this reduced the reported CPU time (which seems to wait for the GPU to finish) for the frame from ~9.5ms to ~4.5ms, and on my laptop (Ubuntu 18.04, Intel GPU, Chromium) from ~25ms to ~6.5ms. Changing the dot shape to square further improves performance (e.g. on desktop, from 4.5ms to ~2.75ms).

I haven't changed any documentation yet because some details still need to be fleshed out:

 - If `dot_shape` + square is desirable, should it share `dot_radius` with the circle shape or should it have a separate parameter to allow smaller shapes/should `dot_radius` be allowed to be fractional?
 - Should available `dot_shapes` match the integers used for `aperture_type`, i.e. 1 for circle and 3 for square?
 - Should code target ES5 or 6+? I tried to match style in some places, but also used `const` + arrow functions...

Hope this is useful!