@jodeleeuw 2022-02-28
Thanks @aforren1! This looks great.

> If dot_shape + square is desirable, should it share dot_radius with the circle shape or should it have a separate parameter to allow smaller shapes/should dot_radius be allowed to be fractional?

I'm not someone who frequently uses this plugin, but it sounds like a desirable change to me. I think `dot_radius` would be confusing as a square parameter if the way that it will render is the length from the center to the corners. I just had to look up the definition of the radius of a square, for example 😁.

> Should available dot_shapes match the integers used for aperture_type, i.e. 1 for circle and 3 for square?

I admit I find the `aperture_type` parameter confusing right now -- in the rest of the library we tend to use descriptive strings for these kinds of parameters. So perhaps `dot_shapes` should be a string?

> Should code target ES5 or 6+? I tried to match style in some places, but also used const + arrow functions...

6+ is great. We're slowly migrating the codebase in that direction. This is just a piece we haven't gotten to yet.